requirements.txt
async-lru==1.0.2
pyaes==1.6.1
Pyrogram @ git+https://github.com/allen0099/pyrogram.git@53d213b5fa48bab31ee4c34cd270df2f52e01614
PySocks==1.7.1
python-dotenv==0.19.2
TgCrypto==1.2.2
